int module init, file renames (#447)
* use proper internal module timer alias from hal.h,
* rename *_arm to generic file names,
* added APB2 SYSCFG define to hal.h, removed from intmoduleAfhds2aStart,
* cleanup of dead code

+++ radio/src/targets/flysky/telemetry_driver.cpp
@@ -140,7 +140,7 @@ void telemetryPortSetDirectionInput() {
   TELEMETRY_DMA_Channel_RX->CCR |= DMA_CCR_EN;
 }
 
-void sportSendBuffer(const uint8_t* buffer, unsigned long count) {
+void sportSendBuffer(const uint8_t* buffer, uint32_t count) {
   telemetryPortSetDirectionOutput();
 
   DMA_DeInit(TELEMETRY_DMA_Channel_TX);
